// src/models/slack-command.ts
export class SlackCommandPostBody {
    token: string;
    command: string;
    text: string;
    response_url: string;
    trigger_id: string;
    user_id: string;
    user_name: string;
    team_id: string;
    team_name: string;
    enterprise_id: string;
    enterprise_name: string;
    channel_id: string;
    channel_name: string;
    api_app_id: string;
}